WIP breaking loadtopics file into multiple files
This file was doing too much, so each loader now has it's own class,
with common function inherited from the BaseLoader.
